+++ src/main/java/MjLee/boardService/controller/UserController.java
@@ -2,6 +2,7 @@ package MjLee.boardService.controller;
 
 import MjLee.boardService.dto.UserDto;
 import MjLee.boardService.service.CommentService;
+import MjLee.boardService.service.LoginService;
 import MjLee.boardService.service.PostingService;
 import MjLee.boardService.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,17 +14,19 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/board/user")
 public class UserController {
     UserService userService;
     PostingService postingService;
     CommentService commentService;
+    LoginService loginService;
 
     @Autowired
-    public UserController(UserService userService, PostingService postingService, CommentService commentService) {
+    public UserController(UserService userService, PostingService postingService, CommentService commentService, LoginService loginService) {
         this.userService = userService;
         this.postingService = postingService;
         this.commentService = commentService;
+        this.loginService = loginService;
     }
 
     @GetMapping("{userName}")
@@ -54,18 +57,34 @@ public class UserController {
         return ResponseEntity.status(HttpStatus.OK).build();
     }
 
-    @DeleteMapping("{userName}/{userPassWord}") // login을 구현하면 password를 따로 받을 필요 없을 것 같음
-    public ResponseEntity<Void> deleteUser(@PathVariable String userName, @PathVariable String userPassWord){
+    @PutMapping("/login")
+    public ResponseEntity<Void> login(@RequestBody UserDto userDto){
+        if(loginService.login(userDto)) return ResponseEntity.status(HttpStatus.OK).build();
+        else return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+    }
+
+    @PutMapping("/logout")
+    public ResponseEntity<Void> logout(@RequestBody UserDto userDto){
+        try{
+            loginService.logout(userDto);
+        } catch (RuntimeException e){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).build();
+    }
+
+    @DeleteMapping("{userName}")
+    public ResponseEntity<Void> deleteUser(@PathVariable String userName){
+
         UserDto userDto = new UserDto();
         userDto.setUserName(userName);
-        userDto.setPassword(userPassWord);
+        if(userService.findUser(userDto).isLogin()){
+            commentService.deleteUser(userName);
+            postingService.deleteUser(userName);
 
-        commentService.deleteUser(userName);
-        postingService.deleteUser(userName);
-
-
-        if(userService.delete(userDto)) return ResponseEntity.status(HttpStatus.OK).build();
-
+            if(userService.delete(userDto)) return ResponseEntity.status(HttpStatus.OK).build();
+        }
 
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
     }
